do refactor for remote config
