+++ component/notify/componet_notify_test.go
@@ -19,12 +19,13 @@ package notify
 
 import (
 	"encoding/json"
-	"fmt"
 	"github.com/zouyx/agollo/v4/cluster/roundrobin"
+	"github.com/zouyx/agollo/v4/component/log"
+	"github.com/zouyx/agollo/v4/constant"
 	jsonFile "github.com/zouyx/agollo/v4/env/file/json"
-	"net/http"
-	"net/http/httptest"
+	"github.com/zouyx/agollo/v4/utils"
 	"os"
+	"path"
 	"testing"
 	"time"
 
@@ -45,89 +46,6 @@ func init() {
 var (
 	jsonConfigFile = &jsonConfig.ConfigFile{}
 )
-
-const responseStr = `[{"namespaceName":"application","notificationId":%d}]`
-
-func onlyNormalConfigResponse(rw http.ResponseWriter, req *http.Request) {
-	rw.WriteHeader(http.StatusOK)
-	fmt.Fprintf(rw, configResponseStr)
-}
-
-func onlyNormalTwoConfigResponse(rw http.ResponseWriter, req *http.Request) {
-	rw.WriteHeader(http.StatusOK)
-	fmt.Fprintf(rw, configAbc1ResponseStr)
-}
-
-func onlyNormalResponse(rw http.ResponseWriter, req *http.Request) {
-	result := fmt.Sprintf(responseStr, 3)
-	fmt.Fprintf(rw, "%s", result)
-}
-
-func initMockNotifyAndConfigServer() *httptest.Server {
-	//clear
-	handlerMap := make(map[string]func(http.ResponseWriter, *http.Request), 1)
-	handlerMap["application"] = onlyNormalConfigResponse
-	handlerMap["abc1"] = onlyNormalTwoConfigResponse
-	return runMockConfigServer(handlerMap, onlyNormalResponse)
-}
-
-func TestSyncConfigServices(t *testing.T) {
-	server := initMockNotifyAndConfigServer()
-	appConfig := initNotifications()
-	appConfig.IP = server.URL
-	apolloConfigs := AsyncConfigs(appConfig)
-	//err keep nil
-	Assert(t, apolloConfigs, NotNilVal())
-	Assert(t, len(apolloConfigs), Equal(2))
-}
-
-func TestGetRemoteConfig(t *testing.T) {
-	server := initMockNotifyAndConfigServer()
-
-	time.Sleep(1 * time.Second)
-
-	var remoteConfigs []*config.Notification
-	var err error
-	appConfig := initNotifications()
-	appConfig.IP = server.URL
-	remoteConfigs, err = notifyRemoteConfig(appConfig, EMPTY)
-
-	//err keep nil
-	Assert(t, err, NilVal())
-
-	Assert(t, remoteConfigs, NotNilVal())
-	Assert(t, 1, Equal(len(remoteConfigs)))
-	t.Log("remoteConfigs:", remoteConfigs)
-	t.Log("remoteConfigs size:", len(remoteConfigs))
-
-	notify := remoteConfigs[0]
-
-	Assert(t, "application", Equal(notify.NamespaceName))
-	Assert(t, true, Equal(notify.NotificationID > 0))
-}
-
-func TestErrorGetRemoteConfig(t *testing.T) {
-	//clear
-	initNotifications()
-	appConfig := initNotifications()
-	server := runErrorResponse()
-	appConfig.IP = server.URL
-	appConfig.NextTryConnTime = 0
-
-	time.Sleep(1 * time.Second)
-
-	var remoteConfigs []*config.Notification
-	var err error
-	remoteConfigs, err = notifyRemoteConfig(appConfig, EMPTY)
-
-	Assert(t, err, NotNilVal())
-	Assert(t, remoteConfigs, NilVal())
-	Assert(t, 0, Equal(len(remoteConfigs)))
-	t.Log("remoteConfigs:", remoteConfigs)
-	t.Log("remoteConfigs size:", len(remoteConfigs))
-
-	Assert(t, "over Max Retry Still Error", Equal(err.Error()))
-}
 
 func initNotifications() *config.AppConfig {
 	appConfig := env.InitFileConfig()
@@ -174,13 +92,6 @@ func TestUpdateAllNotificationsError(t *testing.T) {
 	appConfig.GetNotificationsMap().UpdateAllNotifications(notifies)
 
 	Assert(t, appConfig.GetNotificationsMap().GetNotifyLen(), Equal(2))
-}
-
-func TestToApolloConfigError(t *testing.T) {
-
-	notified, err := toApolloConfig([]byte("jaskldfjaskl"))
-	Assert(t, notified, NilVal())
-	Assert(t, err, NotNilVal())
 }
 
 func TestAutoSyncConfigServices(t *testing.T) {
@@ -263,39 +174,4 @@ func getTestAppConfig() *config.AppConfig {
 	appConfig := c.(*config.AppConfig)
 	appConfig.Init()
 	return appConfig
-}
-
-func TestCreateApolloConfigWithJson(t *testing.T) {
-	jsonStr := `{
-  "appId": "100004458",
-  "cluster": "default",
-  "namespaceName": "application",
-  "configurations": {
-    "key1":"value1",
-    "key2":"value2"
-  },
-  "releaseKey": "20170430092936-dee2d58e74515ff3"
-}`
-
-	config, err := createApolloConfigWithJSON([]byte(jsonStr))
-
-	Assert(t, err, NilVal())
-	Assert(t, config, NotNilVal())
-
-	Assert(t, "100004458", Equal(config.AppID))
-	Assert(t, "default", Equal(config.Cluster))
-	Assert(t, "application", Equal(config.NamespaceName))
-	Assert(t, "20170430092936-dee2d58e74515ff3", Equal(config.ReleaseKey))
-	Assert(t, "value1", Equal(config.Configurations["key1"]))
-	Assert(t, "value2", Equal(config.Configurations["key2"]))
-
-}
-
-func TestCreateApolloConfigWithJsonError(t *testing.T) {
-	jsonStr := `jklasdjflasjdfa`
-
-	config, err := createApolloConfigWithJSON([]byte(jsonStr))
-
-	Assert(t, err, NotNilVal())
-	Assert(t, config, NilVal())
 }
