+++ component/notify/config_server_test.go
@@ -96,16 +96,6 @@ func runNormalConfigResponse() *httptest.Server {
 
 //Error response
 //will hold 5s and keep response 404
-func runErrorResponse() *httptest.Server {
-	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.WriteHeader(http.StatusNotFound)
-	}))
-
-	return ts
-}
-
-//Error response
-//will hold 5s and keep response 404
 func runErrorConfigResponse() *httptest.Server {
 	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		time.Sleep(500 * time.Microsecond)
