+++ component/remote/sync/sync_test.go
@@ -15,14 +15,4 @@
  * limitations under the License.
  */
 
-package component
-
-//AbsComponent 定时组件
-type AbsComponent interface {
-	Start()
-}
-
-//StartRefreshConfig 开始定时服务
-func StartRefreshConfig(component AbsComponent) {
-	component.Start()
-}
+package sync
